Add Lettuce Redis watcher support to Casbin AutoConfiguration
The LettuceRedisWatcher class is imported, and a conditional statement is added to check the watcher type. If the type is "standalone", a new LettuceRedisWatcher instance is created and set as the watcher. If the type is "cluster", a new LettuceRedisWatcher instance is created with the cluster nodes and set as the watcher. If the watcher type is unsupported, a CasbinWatcherLettuceTypeUnsupportedException is thrown.