feat: add Lettuce Redis watcher (#89)
* Update dependencies and configuration for Casbin:
- Update `org.casbin:jcasbin` from version 1.30.1 to 1.35.0
- Update `org.casbin:jcasbin-redis-watcher` from version 1.4.1 to 1.4.2
- Update `org.casbin:jdbc-adapter` from version 2.3.3 to 2.4.1
- Update `mysql:mysql-connector-java` from version 8.0.30 to 8.0.33

Add new configuration properties for Lettuce Redis Type in `CasbinProperties.java`:
- Add `watcherLettuceRedisType` property with the default value of NONE

Update getter and setter methods for `watcherLettuceRedisType` in `CasbinProperties.java`

Add new class `CasbinWatcherLettuceRedisType.java`

Update `CasbinRedisWatcherAutoConfiguration.java`:
- Add import statement for `org.casbin.watcher.lettuce.LettuceRedisWatcher`
- Add new conditional bean configuration for Lettuce Redis Watcher

* Refactor build.gradle dependencies and update Casbin classes

* Add Lettuce Redis watcher support to Casbin AutoConfiguration

The LettuceRedisWatcher class is imported, and a conditional statement is added to check the watcher type. If the type is "standalone", a new LettuceRedisWatcher instance is created and set as the watcher. If the type is "cluster", a new LettuceRedisWatcher instance is created with the cluster nodes and set as the watcher. If the watcher type is unsupported, a CasbinWatcherLettuceTypeUnsupportedException is thrown.

* Update dependencies version and fix unsupported watcher type exception in CasbinRedisWatcherAutoConfiguration.java.